Fix some broken specs for TimeSeries
Apparently RedisTimeSeries clamped down on some of the requirements for
chunks and changed how it emits some of the `nan` values (now `-nan`),
so we need to update the specs to match what the server returns.

I don't love this since it won't run the expectations correctly against
older RedisTimeSeries versions on the server, but this is a pragmatic
change for now.