Add LMOVE
Since LMOVE takes two low-cardinality arguments, they're implemented as
enums. Enum#to_s involves no allocation (the strings are pre-allocated),
so this remains fast, but because `override_return_types` was doing
simple delegation, this meant that type inference for symbol->enum
wasn't working, so this commit changes that up to use the original type
signature of the command method.