RedisGraph has been removed from Redis Stack
https://redis.com/blog/redisgraph-eol/

Honestly, this was the weirdest part of this driver. But as much as I'm
glad that I might not need to maintain this, I'm kinda sad to see it go.
Graph DBs are soooo good — probably the single best database paradigm in
existence, at least in how it matches the way people often think about
their data. Just, ugh.

But since Redis Stack has removed it, we can't even test this driver
with it anymore.