Return errors in the parser rather than raising
The protocol parser isn't the right layer to be raising exceptions.
Instead, we let the connection, which is interpreting the parsed values,
decide what to do with the error returned from the parser.

This is necessary to handle errors raised inside of a MULTI/EXEC
transaction. We have to be able to read the results from *all* commands
executed inside the transaction, we can't raise halfway through for a
Redis error.