Fix scan_each on Dragonfly DB
An empty cursor value is valid on Redis, but DragonflyDB requires it.
The cursor was an empty string so that it wouldn't be set to the
terminal value for the loop, which is also supposed to be the starting
value. Redis accepted the empty string as an equivalent value, so I just
ran with that, but other implementations seem to be a bit more strict on
some semantics.