Avoid partial reads in Simple String optimization
This can happen in long pipelines, which is really what this
optimization was intended to help. If we're writing large command
pipelines (such as seeding a Redis DB, warming the cache for large
quantities of documents, etc), we need to ensure that simple strings
that are returned are read fully so that we don't mess up the alignment
in the parser.