Avoid raising while committing pipelines
For a pipeline to function properly, we need to have a 1:1 mapping of
commands to responses. If we raise while committing the pipeline, we
don't read any of the responses after the first error response. So if we
sent 10 commands and the 5th response is an error, it was not reading
the responses of commands 6-10, leaving them in the TCP buffer. This
results in the next command sent (we'll call it command #11) reading the
result of command #6, which is the wrong response.

So we need to *return* the results of all of the commands on the
pipeline rather than raising error responses as errors. This is also
important because there could be multiple errors in the pipeline, so we
shouldn't just raise the first one we see and ignore the rest.