Clean up XADD command
The keyword-argument form of `xadd` has been painful since its
inception. I thought it would be so nice, but it's just turned out to
cause headaches for maintainability. It would occasionally cause
arguments like `maxlen` to be seen as event properties, which would
silently do two wrong things:

- Not cap the stream
- Add an unnecessary property to the event

I was actually wondering the other day why Redis stream events are
enforced as field/value pairs instead of an opaque payload. It's not
like it allows you to filter on the event properties, so it seems like
unnecessary ceremony. The reason I mention this is that this confusion
never would've happened if Redis just treated events as opaque instead
of field/value pairs.