Handle READONLY errors by reinitializing
If replication topology goes by DNS, this may be all that is needed to
self-heal. This won't be a fully self-healing solution for connections
going directly to an IP address, though. We'll need to come up with
another solution for that.

This commit also adds a `closed?` method, which is how the connection
pool knows to discard the connection.