Handle server disconnections better in Client
The `Client` now closes a connection if it cannot self-heal. This will
cause `DB::Pool` to discard it and spin up a new one if needed.

Additionally, `Connection#close` now ignores errors. The socket will try
to flush before closing, but if it can't talk to the server it raises an
error. If it can't flush, we need to close the connection anyway, so we
ignore the exception. This is a common operation in a lot of database
clients I've found.