Distinguish between Immediate and Deferred command
An "immediate" command is one that is sent to the server immediately and
its result can be read immediately. This is the most common form of
operation. The `Client`, `Connection`, `Cluster`, and
`ReplicationClient` all use this paradigm.

A "deferred" command is one where the results are not available
immediately. The two main types that use this concept are `Pipeline` and
`Transaction`.