Only run RedisGraph specs when enabled
The RedisGraph module is no longer maintained. It has been extracted
into FalkorDB.

    https://www.falkordb.com

As a result, `Redis::Graph` has also been extracted from this client to
a `FalkorDB` shard:

    https://github.com/jgaskins/falkordb

Support for RedisGraph will remain in this client until I feel like removing it, but there'll be a deprecation in place.