Allow MULTI blocks to return or break
Prior to this commit, if a `multi` block had a `return` or `break` in
it, the `exec` or `discard` would never be executed, leaving the
connection with an open transaction. The consequences of this are:
- the client and server disagree about the transaction state
- the server will not execute any further commands on that connection
  until it receives an `EXEC` command
  - this results in infinite accrual of queued commands
- every command sent to the server will only result in returning the
  string "QUEUED" until the server receives an `EXEC` command
- the connection cannot run any additional transactions because Redis
  transactions cannot be nested
- if, for some strange reason, the server does receive an `EXEC` on this
  connection, however many commands have been enqueued will immediately
  execute, blocking the server until they are completed

This commit ensures that the connection runs `EXEC` if the transaction
block contains a `return` or `break` by checking if the transaction is
in a `queued` state and running `EXEC` if so.